Removed entities
