--- src/Persistence.Abstractions/Cell/ICellEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace GoodToCode.Persistence.Abstractions
-{
-    public interface ICellEntity : ICellData, IEntity
-    {
-
-    }
-}
