+++ src/Persistence.Abstractions/Persistence.Abstractions.csproj
@@ -36,4 +36,8 @@
     <PackageReference Include="Microsoft.Extensions.Options" Version="6.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="IEntity\" />
+  </ItemGroup>
+
 </Project>
