--- src/Persistence.Abstractions/Row/IRowEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace GoodToCode.Persistence.Abstractions
-{
-    public interface IRowEntity : IRowData, IEntity
-    {
-
-    }
-}
