--- src/Persistence.Abstractions/Services/IEntityPersistenceService.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace GoodToCode.Persistence.Abstractions
-{
-    public interface IEntityPersistenceService<T> where T : IEntity, new()
-    {
-        Task<IEnumerable<T>> GetItemsAsync(string query);
-        Task<T> GetItemAsync(string id, string partitionKeyPath);
-        Task AddItemAsync(T item);
-        Task UpdateItemAsync(string id, T item);
-        Task DeleteItemAsync(string id, string partitionKeyPath);
-    }
-}
