--- src/Persistence.Abstractions/Sheet/ISheetEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace GoodToCode.Persistence.Abstractions
-{
-    public interface ISheetEntity : ISheetData, IEntity
-    {
-
-    }
-}
