--- src/Persistence.Abstractions/Workbook/IWorkbookEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace GoodToCode.Persistence.Abstractions
-{
-    public interface IWorkbookEntity : IWorkbookData, IEntity
-    {
-
-    }
-}
