+++ src/Persistence.Tests/Factories/RowDataFactory.cs
@@ -1,0 +1,15 @@
+﻿using GoodToCode.Persistence.Abstractions;
+using System;
+using System.Collections.Generic;
+
+namespace GoodToCode.Persistence.Tests
+{
+    public class RowDataFactory
+    {
+        public static RowData CreateRowData()
+        {
+            var cells = new List<CellData>() { CellDataFactory.CreateCellData() };
+            return new RowData(0, cells);
+        }
+    }
+}
