+++ src/Persistence.Tests/Factories/WorkbookDataFactory.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 
 namespace GoodToCode.Persistence.Tests
 {
-    public class WorkbookFactory
+    public class WorkbookDataFactory
     {
         public static WorkbookData CreateWorkbookData()
         {
-            var sheets = new List<SheetData>() { SheetFactory.CreateSheetData() };
+            var sheets = new List<SheetData>() { SheetDataFactory.CreateSheetData() };
             return new WorkbookData("book1.xls", sheets);
         }
     }
