+++ src/Persistence.Tests/Workbook/Excel_Workbook_Tests.cs
@@ -17,10 +17,7 @@ namespace GoodToCode.Persistence.Tests
     {
         private readonly ILogger<Excel_Workbook_Tests> logItem;
         private static string SutXlsxFile { get { return @$"{PathFactory.GetProjectSubfolder("Assets")}/OpinionFile.xlsx"; } }
-        public CellEntity SutRow { get; private set; }
-        public IEnumerable<CellEntity> SutRows { get; private set; }
         public Dictionary<string, StringValues> SutReturn { get; private set; }
-
 
         public Excel_Workbook_Tests()
         {
